+++ src/test/java/org/team6/coffeebeanery/order/controller/SellerOrderControllerTest.java
@@ -110,7 +110,7 @@ public class SellerOrderControllerTest {
     public List<Order> createOrders() throws InterruptedException {
         for (int i = 1; i <= 17; i++) {
             Delivery delivery = new Delivery();
-            delivery.setDeliveryNumber((double) i * 1000);
+            delivery.setDeliveryNumber((long) i * 1000);
             delivery.setDeliveryCompany("우체국");
 
             List<OrderDetail> orderDetails = new ArrayList<>();
@@ -122,7 +122,7 @@ public class SellerOrderControllerTest {
                     .address(new Address("baseAddress" + i,
                             "detailAddress" + i,
                             "z" + (int) i))
-                    .totalPrice((double) i * 10000)
+                    .totalPrice((long) i * 10000)
                     .orderStatus(OrderStatus.ORDERED)
                     .delivery(delivery)
                     .orderDetails(orderDetails)
